AI Model/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
scikit-image
pydantic